requirements.txt
streamlit~=1.40.1
requests~=2.32.3
torch~=2.4.1
transformers~=4.46.3
pillow~=10.4.0